+++ src/cpu/cpu.c
@@ -298,7 +298,7 @@ CPU          *cpu_s;
 uint8_t do_translate  = 0;
 uint8_t do_translate2 = 0;
 
-void (*cpu_exec)(int64_t cycs);
+void (*cpu_exec)(int32_t cycs);
 
 uint8_t ccr0;
 uint8_t ccr1;
